requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-pptx==0.6.23
pydantic==2.5.3
python-multipart==0.0.6